requirements.txt
# ============================================================
# Secure FinTech API — Python Dependencies
# All versions are pinned for reproducible builds.
# Run: pip install -r requirements.txt
# ============================================================

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database — async driver for non-blocking I/O
sqlalchemy==2.0.35
asyncpg==0.29.0          # async PostgreSQL driver
alembic==1.13.3          # database migrations

# Auth & Security
python-jose[cryptography]==3.3.0   # JWT encoding/decoding
passlib[bcrypt]==1.7.4             # password hashing
python-multipart==0.0.12           # required for OAuth2 form parsing

# Config — reads from environment variables
pydantic==2.9.2
pydantic-settings==2.5.2

# HTTP client (used in tests and health checks)
httpx==0.27.2
